Update ALU.hdl
Xor is a bit of overkill for this function being composed of 4 gates instead of one. You save a variable using not. Also if you would use a more way Xor, the result could be wrong. 

BTW struggled a lot with my own implementation. Used all kinds of demultiplexers and stuff, your version is way better. Thanks for being there.